Move layout and suffix types to cache package
This is the only user of these types
